# Use swc to minify the javascript a lot quicker