--- a/_scripts/webpack.main.config.js
+++ b/_scripts/webpack.main.config.js
@@ -2,6 +2,7 @@ const path = require('path')
 const webpack = require('webpack')
 const CopyWebpackPlugin = require('copy-webpack-plugin')
 const JsonMinimizerPlugin = require('json-minimizer-webpack-plugin')
+const TerserPlugin = require('terser-webpack-plugin')
 
 const { productName } = require('../package.json')
 
@@ -26,7 +27,15 @@ const config = {
   // webpack defaults to only optimising the production builds, so having this here is fine
   optimization: {
     minimizer: [
-      '...', // extend webpack's list instead of overwriting it
+      new TerserPlugin({
+        // swc is significantly faster than terser
+        minify: TerserPlugin.swcMinify,
+        terserOptions: {
+          compress: {
+            hoist_funs: true
+          }
+        }
+      }),
       new JsonMinimizerPlugin({
         exclude: /\/locales\/.*\.json/
       })