# Use swc to minify the javascript a lot quicker

<!-- Thanks for sending a pull request! Make sure to follow the contributing guidelines. -->
<!-- Important note, we may remove your pull request if you do not use this provided PR template correctly. -->

## Pull Request Type

- [x] Other - Build improvement

## Description
While using webpack's ProfilingPlugin to look into what is making our webpack builds take a while, I discovered that the JavaScript minification step with terser is taking up almost half of the build time. Please note that the ProfilingPlugin slows down the build by a few seconds, so the actual second values in the screenshots of the profile, should be taken with a grain of salt, it's still good as a relative comparision though.

I found out that the TerserWebpackPlugin that webpack uses also supports other minifiers than terser, like swc (Speedy Web Compiler) and esbuild. I tried both swc and esbuild, with a bit of tweaking I was able to get the swc output to be only a few kbs larger for the renderer.js file and less than a kb larger for the main.js file, being significantly faster in the process:

(I used `yarn run pack` instead of `yarn build` for these measurements, so that we would only get the timings for the webpack step, without electron-builder.)

`yarn run pack` before:

```
...
main (webpack 5.75.0) compiled successfully in 2063 ms
...
renderer (webpack 5.75.0) compiled successfully in 17473 ms
Done in 19.48s.
```

`yarn run pack` after:

```
...
main (webpack 5.75.0) compiled successfully in 1468 ms
...
renderer (webpack 5.75.0) compiled successfully in 9769 ms
Done in 11.83s.
```

The timings were all measured on my computer, so the numbers are already a lot lower than the CI build, however these changes should also make a difference to the CI build.

**I didn't make this change for the web build, as when I tried it, the output increased by almost a mb, which was too much for my liking.**

https://webpack.js.org/plugins/terser-webpack-plugin/#swc

## Screenshots <!-- If appropriate -->
### Profiling:
before:
![profile_before](https://user-images.githubusercontent.com/48293849/205992626-0240ec34-f69b-4dd2-8fe7-42716cc5a340.png)

after:
![profile_after](https://user-images.githubusercontent.com/48293849/205992639-4f1dc8c9-5300-4b09-b1b7-6e21cb6779eb.png)

### renderer.js:
before:
![renderer_before](https://user-images.githubusercontent.com/48293849/205992840-5a01b798-ada6-456e-adc0-68f4033cd9a9.png)

after:
![renderer_after](https://user-images.githubusercontent.com/48293849/205993060-d68c2b99-71d5-420e-8921-a6524bd434d5.png)


### main.js
before:
![main_before](https://user-images.githubusercontent.com/48293849/205993091-3a837566-ea34-414c-b778-cb2162956c55.png)

after:
![main_after](https://user-images.githubusercontent.com/48293849/205993141-a1754126-8fa9-4ef8-8d5a-e50027679ce9.png)

### Resources:
before:
![resources_before](https://user-images.githubusercontent.com/48293849/205993201-a7d5d7ec-fdff-42ea-8336-5e72a240284a.png)

after:
![resources_after](https://user-images.githubusercontent.com/48293849/205993240-4176f228-1211-4409-8a85-a0a77d79f453.png)

## Testing <!-- for code that is not small enough to be easily understandable -->
`yarn build` and then test that FreeTube works the same way as it did before.

## Desktop
<!-- Please complete the following information-->
- **OS:** Windows
- **OS Version:** 10
- **FreeTube version:** 0.18.0